+++ examples/ex002.py
@@ -1,9 +1,9 @@
-'''Illustrates usage of Recorder and Method as RecordableDevices
+'''Illustrates non-decorator usage.
 '''
 
 import os
 import sys
-from rekorder import Recorder, When
+from rekorder import Recorder, When, What
 
 
 def main():
@@ -13,18 +13,11 @@ def main():
 if __name__ == "__main__":
 
   # Get a named Recorder instance.
-  recorder = Recorder.get_recorder(name=__name__, output=os.path.basename(sys.argv[0]).replace('.py', '.json'))
+  recorder = Recorder.get_recorder(
+      name=os.path.basename(__file__).replace('.py', ''),
+      output=os.path.basename(__file__).replace('.py', '.json'))
 
-  # The replayable recording doesn't start until the @recorder.begin
-  # function is invoked. Any recorded information prior to that will
-  # be used as playback configuration.
-
-  if len(sys.argv) > 1:
-    # Capture the state of one or more repositories in the recording's
-    # header (configuration) section.
-    # On playback the repositories will be cloned and checked out at
-    # the recorded commit.
-    recorder.repository_state.record(paths=sys.argv[1:])
+  print("Application recorder is [{}]".format(recorder))
 
   # Instead of using decorators as the other examples do, we can "manually"
   # decorate the function by invoking the decorator functions.
@@ -35,13 +28,17 @@ if __name__ == "__main__":
 
   main()
 
-  # The recordable recording stops when the @recorder.end function
-  # is invoked. Anything recorded after that will be used to restore
-  # the workspace during playback.
+else:  # If we _don't_ want to do playback verification, we _don't_ need this.
 
-  if len(sys.argv) > 1:
-    # Capture the state of those same repositories in the recording's
-    # trailter (restore) section.
-    # On playback the repositories will be checked out at the recorded
-    # commit. (i.e. - restored to their pre-execution state)
-    recorder.repository_state.record(paths=sys.argv[1:])
+  # CLI apps generally don't have an `else` for their check against __name__.
+  # However, for playback verification to work, we need to wrap main() in this
+  # state just as we did when recording. If we don't do this, the tracks will
+  # be out of sync and verification will fail.
+
+  # Get a named Recorder instance.
+  recorder = Recorder.get_recorder(
+      name=os.path.basename(__file__).replace('.py', ''),
+      output=os.path.basename(__file__).replace('.py', '.json'))
+
+  main = recorder.begin(main)
+  main = recorder.end(main)