+++ examples/ex003.py
@@ -1,4 +1,4 @@
-'''Illustrates basic rekorder usage with nested methods and skip.
+'''Illustrates basic rekorder usage with nested methods.
 '''
 
 import os
@@ -16,19 +16,22 @@ recorder = Recorder.get_recorder(
 @recorder.end    # End the recording when main() returns.
 #                # This also captures the method's result.
 def main():
+  print("Begin [{}]".format(__name__))
   foo(1, 2, 3, foo='xxx', baz='yyy')
+  print("End [{}]".format(__name__))
 
 
+@recorder.method.repository
 @recorder.method.params         # Capture foo's parameters before invocation.
 @recorder.method.pass_recorder  # This should be "closest" to the function
 #                               # because a Recorder cannot be serialized.
 def foo(recorder, *args, **kwargs):
 
-  # @recorder.playback.skip  # Not yet implemented
   @recorder.method.params(when=When.AROUND)  # Capture bar's parameters before
   #                                          # and after method invocation.
   def bar():
-    pass
+    if len(sys.argv) > 1:
+      recorder.repository_manager.record(paths=sys.argv[1:])
 
   bar()
 
@@ -43,7 +46,7 @@ if __name__ == "__main__":
     # header (configuration) section.
     # On playback the repositories will be cloned and checked out at
     # the recorded commit.
-    recorder.repository_state.record(paths=sys.argv[1:])
+    recorder.repository_manager.record(paths=sys.argv[1:])
 
   main()
 
@@ -56,4 +59,4 @@ if __name__ == "__main__":
     # trailter (restore) section.
     # On playback the repositories will be checked out at the recorded
     # commit. (i.e. - restored to their pre-execution state)
-    recorder.repository_state.record(paths=sys.argv[1:])
+    recorder.repository_manager.record(paths=sys.argv[1:])