+++ rekorder/lib/method/rval.py
@@ -22,6 +22,8 @@ class MethodReturn(Decorator):
   def __eq__(self, other):
     '''Compare the attributes common between describe_playable_device() and describe_recordable_device()
     '''
+    if type(self) != type(other):
+      return False
     if self.__class__ != other.__class__:
       return False
     if self.when != other.when: